refactor/#126 : 지도 상태관련 로직과 위치관련 로직을 훅으로 문리한다.
